+++ src/index.ts
@@ -1,6 +1,1 @@
-/**
- * Just a simple function,
- */
-export default function todo() {
-  return 'todo';
-}
+export * from './set';