+++ src/set.built.test.ts
@@ -1,0 +1,8 @@
+import { createImportFnTests } from '@bemedev/dev-utils/build-tests';
+
+describe(
+  ...createImportFnTests({
+    FAILS: [],
+    SUCCESS: ['createBetterSet', 'createSet'],
+  }),
+);